+++ gomen/src/main/java/com/tuna/gomen/bookMarkandLike/command/dto/LikeDTO.java
@@ -1,0 +1,15 @@
+package com.tuna.gomen.bookMarkandLike.command.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class LikeDTO {
+    private Integer userId;
+    private Integer postId;
+}
