--- gomen/src/main/java/com/tuna/gomen/bookMarkandLike/dto/BookmarkDTO.java
@@ -1,15 +1,0 @@
-package com.tuna.gomen.bookMarkandLike.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookmarkDTO {
-    private Integer userId;
-    private Integer postId;
-}
