--- gomen/src/main/java/com/tuna/gomen/bookMarkandLike/dto/BookmarkedBoardDto.java
@@ -1,16 +1,0 @@
-package com.tuna.gomen.bookMarkandLike.dto;
-
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class BookmarkedBoardDto {
-    private Integer postId;
-    private String title;
-    private String isBlinded;
-    private String isDeleted;
-    private String userNickname;
-
-}
