+++ gomen/src/main/java/com/tuna/gomen/bookMarkandLike/query/controller/LikeController.java
@@ -1,0 +1,34 @@
+package com.tuna.gomen.bookMarkandLike.query.controller;
+
+import com.tuna.gomen.bookMarkandLike.query.service.LikeService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@RestController
+@RequestMapping("/gomen/like")
+public class LikeController {
+
+    private final LikeService likeService;
+
+    @Autowired
+    public LikeController(LikeService likeService) {
+        this.likeService = likeService;
+    }
+
+    @GetMapping("/{postId}/count")
+    public ResponseEntity<Map<String, Integer>> getLikeCount(@PathVariable int postId) {
+        int likeCount = likeService.getLikeCount(postId);
+        Map<String, Integer> response = new HashMap<>();
+        response.put("likeCount", likeCount);
+        return ResponseEntity.ok(response);
+    }
+
+
+}
