+++ gomen/src/main/java/com/tuna/gomen/mapper/LikeMapper.java
@@ -1,0 +1,8 @@
+package com.tuna.gomen.mapper;
+
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface LikeMapper {
+    int getLikeCountByPostId(int postId);
+}
