+++ gomen/src/main/resources/mapper/LikeMapper.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.tuna.gomen.mapper.LikeMapper">
+
+    <select id="getLikeCountByPostId" resultType="int">
+        SELECT COUNT(*) FROM tbl_like WHERE post_id = #{postId}
+    </select>
+
+</mapper>
